refactor: Divide InputFile to share the same style
