+++ packages/smarthr-ui/src/components/InputFile/style.ts
@@ -1,0 +1,43 @@
+import { tv } from 'tailwind-variants'
+
+export const inputFile = tv({
+  slots: {
+    wrapper: 'smarthr-ui-InputFile shr-block',
+    fileList: ['smarthr-ui-InputFile-fileList', 'shr-list-none shr-self-stretch shr-text-base'],
+    fileItem: 'shr-flex shr-items-center',
+    inputWrapper: [
+      'shr-border-shorthand shr-relative shr-inline-flex shr-rounded-m shr-bg-white shr-font-bold shr-leading-none',
+      'contrast-more:shr-border-high-contrast',
+      'focus-within:shr-focus-indicator',
+      'has-[[aria-invalid]]:shr-border-danger',
+    ],
+    input: [
+      'smarthr-ui-InputFile-input',
+      'shr-absolute shr-left-0 shr-top-0 shr-h-full shr-w-full shr-opacity-0',
+      'file:shr-h-full file:shr-w-full file:shr-cursor-pointer',
+      'file:disabled:shr-cursor-not-allowed',
+    ],
+    prefix: 'shr-me-0.5 shr-inline-flex',
+  },
+  variants: {
+    size: {
+      default: {
+        inputWrapper: 'shr-px-1 shr-py-0.75 shr-text-base',
+      },
+      s: {
+        inputWrapper: 'shr-p-0.5 shr-text-sm',
+      },
+    },
+    disabled: {
+      true: {
+        inputWrapper: 'shr-border-disabled shr-bg-white-darken shr-text-disabled',
+      },
+      false: {
+        inputWrapper: 'hover:shr-border-darken hover:shr-bg-white-darken hover:shr-text-black',
+      },
+    },
+  },
+  defaultVariants: {
+    size: 'default',
+  },
+})
